lastdigit

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -1,10 +1,10 @@
 #include "main.h"
 /**
  * print_last_digit - print last digit
- * @c: character check
+ * @l: character check
  * Return: 1 sucess or 0 in fail
  */
-int print_last_digit(int l)
+int print_last_digit(long int l)
 {
 	if (l < 0)
 	{
@@ -12,5 +12,4 @@ int print_last_digit(int l)
 	}
 	_putchar((l % 10) + 48);
 	return (l % 10);
-	
 }
